# docs/swagger.yaml
basePath: /api
definitions:
  ApiResponse:
    description: Generic API response
    properties:
      data: {}
      responseKey:
        type: string
    type: object
...
  KeywordIn:
    description: KeywordIn represents the input structure for creating a new role.
    properties:
      name:
        description: Name is the name of the role. It is required and should be between
          2 and 40 characters.
        maxLength: 40
        minLength: 2
        type: string
    required:
    - name
    type: object
  KeywordsCount:
    description: KeywordsCount represents the count of keywords.
    properties:
      count:
        description: Count is the number of keywords.
        type: integer
    type: object
  KeywordsDetails:
    description: KeywordsDetails represents detailed information about a specific
      keyword.
    properties:
      createdAt:
        description: CreatedAt is the timestamp indicating when the keyword was created.
        type: string
      id:
        description: ID is the unique identifier for the keyword.
        type: string
      name:
        description: name is the name of the keyword.
        type: string
    type: object
  KeywordsList:
    description: KeywordsList represents a simplified version of the keyword for listing
      purposes.
    properties:
      id:
        description: ID is the unique identifier for the service.
        type: string
      name:
        description: Name is the full name of the service.
        type: string
    type: object
  KeywordsPagination:
    description: KeywordsPagination represents the paginated list of keywords.
    properties:
      items:
        description: Items is a slice containing individual service details.
        items:
          $ref: '#/definitions/ServicesTable'
        type: array
      limit:
        description: Limit is the maximum number of items per page in the pagination.
        type: integer
      page:
        description: Page is the current page number in the pagination.
        type: integer
      totalCount:
        description: TotalCount is the total number of services in the entire list.
        type: integer
    type: object
...
  ServicesTable:
    description: KeywordsTable represents a single service entry in a table.
    properties:
      createdAt:
        description: CreatedAt is the timestamp indicating when the service entry
          was created.
        type: string
      id:
        description: ID is the unique identifier for the user.
        type: string
      name:
        description: name is the  name of the service. It is required and should be
          between 3 and 30 characters.
        maxLength: 30
        minLength: 3
        type: string
    required:
    - name
    type: object
...
host: localhost:8080
info:
  contact:
    email: h.manai@tunisiancloud.com
    name: Hedi Manai
  description: A basic template for managing your project using Golang REST APIs.
  title: Project Name
  version: "0.1"
paths:
...
  /keywords/{serviceID}:
    get:
      description: Get all keywords.
      parameters:
      - description: Page
        in: query
        name: page
        type: integer
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Service ID
        in: path
        name: serviceID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/KeywordsPagination'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/ApiResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ApiResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/ApiResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ApiResponse'
      security:
      - ApiKeyAuth: []
      summary: Get keywords
      tags:
      - Keyword
    post:
      consumes:
      - application/json
      description: Create a new keyword.
      parameters:
      - description: Service ID
        in: path
        name: serviceID
        required: true
        type: string
      - description: Keyword query params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/KeywordIn'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/ApiResponse'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/ApiResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ApiResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/ApiResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ApiResponse'
      security:
      - ApiKeyAuth: []
      summary: Create keyword
      tags:
      - Keyword
  /keywords/{serviceID}/{ID}:
    get:
      description: Get one keyword.
      parameters:
      - description: Service ID
        in: path
        name: serviceID
        required: true
        type: string
      - description: Keyword ID
        in: path
        name: ID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/KeywordsDetails'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/ApiResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ApiResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/ApiResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ApiResponse'
      security:
      - ApiKeyAuth: []
      summary: Get keyword
      tags:
      - Keyword
  /keywords/{serviceID}/count:
    get:
      description: Get number of all keywords.
      parameters:
      - description: Service ID
        in: path
        name: serviceID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/KeywordsCount'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/ApiResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ApiResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/ApiResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ApiResponse'
      security:
      - ApiKeyAuth: []
      summary: Get number of keywords
      tags:
      - Keyword
  /keywords/{serviceID}/list:
    get:
      description: Get list of all keywords.
      parameters:
      - description: Service ID
        in: path
        name: serviceID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/KeywordsList'
            type: array
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/ApiResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ApiResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/ApiResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ApiResponse'
      security:
      - ApiKeyAuth: []
      summary: Get list of  keywords
      tags:
      - Keyword
  /keywords/{serviceId}/{ID}:
    put:
      consumes:
      - application/json
      description: Update keyword.
      parameters:
      - description: Service ID
        in: path
        name: serviceId
        required: true
        type: string
      - description: Keyword ID
        in: path
        name: ID
        required: true
        type: string
      - description: Keyword query params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/KeywordIn'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ApiResponse'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/ApiResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ApiResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/ApiResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ApiResponse'
      security:
      - ApiKeyAuth: []
      summary: Update keyword
      tags:
      - Keyword
...
schemes:
- http
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"